requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow>=10.0.0
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-filter==23.5

# AWS S3 Storage
boto3==1.34.144
django-storages==1.14.2